+++ crates/sui-types/src/error.rs
@@ -934,6 +934,7 @@ impl SuiError {
                     _ => false,
                 }
             }
+            SuiError::Unknown(..) => false,
             // Other variants are assumed to be retriable.
             _ => true,
         }
